refactor: folders and strings
