+++ PokeOntopTest/PokeDetail/Model/PokeListModel.swift
@@ -14,13 +14,6 @@ struct PokeDetail: Identifiable, Codable {
     }
 }
 
-struct PokemonListResponse: Codable {
-    let count: Int
-    let next: String?
-    let previous: String?
-    let results: [PokeDetail]
-}
-
 struct UrlObject: Codable {
     let url: String?
 }
