+++ PokeOntopTest/PokeList/Model/PokemonListResponse.swift
@@ -1,0 +1,9 @@
+
+import Foundation
+
+struct PokemonListResponse: Codable {
+    let count: Int
+    let next: String?
+    let previous: String?
+    let results: [PokeDetail]
+}
