--- PokeOntopTest/Tools/Handlers/ShimmerView.swift
@@ -1,8 +1,0 @@
-//
-//  ShimmerView.swift
-//  PokeOntopTest
-//
-//  Created by Igor Maximo on 29/01/25.
-//
-
-import Foundation
