+++ PokeOntopTest/Tools/Shared/Network/NetworkConstructor.swift
