+++ PokeOntopTest/Tools/Shared/Statics/ApiConstants.swift
