+++ PokeOntopTest/Tools/Shared/Statics/Constants.swift
