+++ PokeOntopTest/Tools/Shared/StringExtension.swift
