+++ PokeOntopTest/en.lproj/Localizable.strings
@@ -3,3 +3,5 @@
 "no_content_available" = "No Content Available!";
 "an_error_occurred" = "An error occurred. Please try again!";
 "try_again" = "Try Again!";
+"evo_chain" = "Evolution Chain";
+"poke_detail" = "Pokemon Detail";
